+++ client/web/src/enterprise/batches/settings/CodeHostSshPublicKey.tsx
@@ -3,6 +3,8 @@ import { noop } from 'lodash'
 import ContentCopyIcon from 'mdi-react/ContentCopyIcon'
 import React, { useCallback, useState } from 'react'
 
+import { TextArea } from '@sourcegraph/wildcard'
+
 import { ExternalServiceKind } from '../../../graphql-operations'
 
 const configInstructionLinks: Record<ExternalServiceKind, string> = {
@@ -53,9 +55,9 @@ export const CodeHostSshPublicKey: React.FunctionComponent<CodeHostSshPublicKeyP
                     </button>
                 )}
             </div>
-            <textarea
+            <TextArea
                 id={LABEL_ID}
-                className="form-control text-monospace mb-3"
+                className="text-monospace mb-3"
                 rows={5}
                 spellCheck="false"
                 value={sshPublicKey}
