+++ client/web/src/enterprise/insights/components/form/repositories-field/RepositoryField.tsx
@@ -1,9 +1,10 @@
 import { Combobox, ComboboxInput, ComboboxPopover } from '@reach/combobox'
 import React, { ChangeEvent, forwardRef, Ref, useImperativeHandle, useRef } from 'react'
 
+import { FlexTextArea } from '@sourcegraph/wildcard'
+
 import { getSanitizedRepositories } from '../../creation-ui-kit/sanitizers/repositories'
 
-import { FlexTextArea } from './components/flex-textarea/FlexTextArea'
 import { SuggestionsPanel } from './components/suggestion-panel/SuggestionPanel'
 import { useRepoSuggestions } from './hooks/use-repo-suggestions'
 import styles from './RepositoriesField.module.scss'
