+++ client/web/src/enterprise/insights/components/insights-view-grid/components/backend-insight/components/drill-down-filters-panel/components/drill-down-filters-form/components/drill-down-reg-exp-input/DrillDownRegExpInput.tsx
@@ -1,10 +1,9 @@
 import classNames from 'classnames'
 import React, { forwardRef, InputHTMLAttributes, PropsWithChildren, Ref } from 'react'
 
-import { Button } from '@sourcegraph/wildcard'
+import { Button, FlexTextArea } from '@sourcegraph/wildcard'
 
 import { TruncatedText } from '../../../../../../../../../../pages/dashboards/dashboard-page/components/dashboard-select/components/trancated-text/TrancatedText'
-import { FlexTextArea } from '../../../../../../../../../form/repositories-field/components/flex-textarea/FlexTextArea'
 
 import styles from './DrillDownRegExpInput.module.scss'
