+++ client/web/src/marketing/SurveyForm.tsx
@@ -4,7 +4,7 @@ import { useHistory } from 'react-router'
 import { Form } from '@sourcegraph/branded/src/components/Form'
 import { useMutation } from '@sourcegraph/shared/src/graphql/apollo'
 import { gql } from '@sourcegraph/shared/src/graphql/graphql'
-import { Button, LoadingSpinner } from '@sourcegraph/wildcard'
+import { Button, LoadingSpinner, TextArea } from '@sourcegraph/wildcard'
 
 import { AuthenticatedUser } from '../auth'
 import { SubmitSurveyResult, SubmitSurveyVariables } from '../graphql-operations'
@@ -106,28 +106,26 @@ export const SurveyForm: React.FunctionComponent<SurveyFormProps> = ({ authentic
                 </div>
             )}
             <div className="form-group">
-                <label className={styles.label} htmlFor="survey-form-score-reason">
-                    What is the most important reason for the score you gave Sourcegraph?
-                </label>
-                <textarea
+                <TextArea
                     id="survey-form-score-reason"
-                    className="form-control"
                     onChange={event => setReason(event.target.value)}
                     value={reason}
                     disabled={response.loading}
                     autoFocus={true}
+                    label={
+                        <span className={styles.label}>
+                            What is the most important reason for the score you gave Sourcegraph?
+                        </span>
+                    }
                 />
             </div>
             <div className="form-group">
-                <label className={styles.label} htmlFor="survey-form-better-product">
-                    What could Sourcegraph do to provide a better product?
-                </label>
-                <textarea
+                <TextArea
                     id="survey-form-better-product"
-                    className="form-control"
                     onChange={event => setBetterProduct(event.target.value)}
                     value={betterProduct}
                     disabled={response.loading}
+                    label={<span className={styles.label}>What could Sourcegraph do to provide a better product?</span>}
                 />
             </div>
             <div className="form-group">
