+++ client/wildcard/src/components/Form/FlexTextArea/index.ts
@@ -1,0 +1,1 @@
+export * from './FlexTextArea'
